server/infra/initial_setup/pyproject.toml
[tool.poetry]
name = "aces-platform-algorithm-server-infra-template"
version = "0.1.0"
description = ""
authors = ["Takaaki Inada <takaaki.inada@acesinc.co.jp>"]

[tool.poetry.dependencies]
python = "^3.8"
boto3 = "^1.17.68"
black = "^21.5b0"
PyYAML = "^5.4.1"
"aws-cdk.core" = "^1.180.0"
"aws-cdk.aws-batch" = "^1.180.0"
"aws-cdk.aws-codepipeline" = "^1.180.0"
"aws-cdk.aws-codepipeline-actions" = "^1.180.0"
"aws-cdk.aws-codebuild" = "^1.180.0"
"aws-cdk.aws-codedeploy" = "^1.180.0"
"aws-cdk.aws-dynamodb" = "^1.180.0"
"aws-cdk.aws-ec2" = "^1.180.0"
"aws-cdk.aws-ecr" = "^1.180.0"
"aws-cdk.aws-ecs" = "^1.180.0"
"aws-cdk.aws-iam" = "^1.180.0"
"aws-cdk.aws-lambda" = "^1.180.0"
"aws-cdk.aws_lambda_event_sources" = "^1.180.0"
"aws-cdk.aws-s3" = "^1.180.0"
"aws-cdk.aws-sns" = "^1.180.0"
"aws-cdk.aws-ssm" = "^1.180.0"
"aws-cdk.custom-resources" = "^1.180.0"
"aws-cdk.aws_xray" = "^1.180.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
